# unsigned: optimizations to improve performance