--- a/zuint32/zuint32.go
+++ b/zuint32/zuint32.go
@@ -42,17 +42,15 @@ func SortBYOB(x, buffer []uint32) {
 
 	from := x
 	to := buffer[:len(x)]
-	var key uint8 // Current byte value
 
 	for keyOffset := uint(0); keyOffset < bitSize; keyOffset += radix {
-		keyMask := uint32(0xFF << keyOffset)
 		var offset [256]int // Keep track of where room is made for byte groups in the buffer
 		sorted := false
 		prev := uint32(0)
 
 		for _, elem := range from {
 			// For each elem to sort, fetch the byte at current radix
-			key = uint8((elem & keyMask) >> keyOffset)
+			key := uint8(elem >> keyOffset)
 			// inc count of bytes of this type
 			offset[key]++
 
@@ -78,7 +76,7 @@ func SortBYOB(x, buffer []uint32) {
 
 		// Swap values between the buffers by radix
 		for _, elem := range from {
-			key = uint8((elem & keyMask) >> keyOffset) // Get the byte of each element at the radix
+			key := uint8(elem >> keyOffset) // Get the byte of each element at the radix
 			to[offset[key]] = elem                     // Copy the element depending on byte offsets
 			offset[key]++                              // One less space, move the offset
 		}