--- a/zuint64/zuint64.go
+++ b/zuint64/zuint64.go
@@ -44,16 +44,15 @@ func SortBYOB(x, buffer []uint64) {
 	from := x
 	to := buffer[:len(x)]
 	var key uint8
-	var offset [256]int // Keep track of where groups start
 
 	for keyOffset := uint(0); keyOffset < bitSize; keyOffset += radix {
 		keyMask := uint64(0xFF << keyOffset) // Current 'digit' to look at
-		var counts [256]int                  // Keep track of the number of elements for each kind of byte
+		var offset [256]int                  // Keep track of where groups start
 		sorted := true                       // Check for already sorted
 		prev := uint64(0)                    // if elem is always >= prev it is already sorted
 		for _, elem := range from {
 			key = uint8((elem & keyMask) >> keyOffset) // fetch the byte at current 'digit'
-			counts[key]++                              // count of elems to put in this digit's bucket
+			offset[key]++                              // count of elems to put in this digit's bucket
 
 			if sorted { // Detect sorted
 				sorted = elem >= prev
@@ -69,9 +68,10 @@ func SortBYOB(x, buffer []uint64) {
 		}
 
 		// Find target bucket offsets
-		offset[0] = 0
-		for i := 1; i < len(offset); i++ {
-			offset[i] = offset[i-1] + counts[i-1]
+		watermark := offset[0] - offset[0] // Like := 0, but inherits the type.
+		for i, count := range offset {
+			offset[i] = watermark
+			watermark += count
 		}
 
 		// Rebucket while copying to other buffer