--- a/zuint64/zuint64.go
+++ b/zuint64/zuint64.go
@@ -43,15 +43,13 @@ func SortBYOB(x, buffer []uint64) {
 	// Each pass processes a byte offset, copying back and forth between slices
 	from := x
 	to := buffer[:len(x)]
-	var key uint8
 
 	for keyOffset := uint(0); keyOffset < bitSize; keyOffset += radix {
-		keyMask := uint64(0xFF << keyOffset) // Current 'digit' to look at
 		var offset [256]int                  // Keep track of where groups start
 		sorted := true                       // Check for already sorted
 		prev := uint64(0)                    // if elem is always >= prev it is already sorted
 		for _, elem := range from {
-			key = uint8((elem & keyMask) >> keyOffset) // fetch the byte at current 'digit'
+			key := uint8(elem >> keyOffset) // fetch the byte at current 'digit'
 			offset[key]++                              // count of elems to put in this digit's bucket
 
 			if sorted { // Detect sorted
@@ -76,7 +74,7 @@ func SortBYOB(x, buffer []uint64) {
 
 		// Rebucket while copying to other buffer
 		for _, elem := range from {
-			key = uint8((elem & keyMask) >> keyOffset) // Get the digit
+			key := uint8(elem >> keyOffset) // Get the digit
 			to[offset[key]] = elem                     // Copy the element to the digit's bucket
 			offset[key]++                              // One less space, move the offset
 		}