The changes are 'trivial' in that they don't change any algorithms or structure of the code.

```console
name               old time/op  new time/op  delta
ZSortUint64T-4     2.71µs ± 7%  2.70µs ±15%     ~     (p=0.684 n=10+10)
ZSorterUint64T-4   2.77µs ± 9%  2.69µs ± 5%     ~     (p=0.243 n=10+9)
ZSortUint64S-4     15.2µs ± 8%  13.2µs ± 5%  -12.98%  (p=0.000 n=10+10)
ZSorterUint64S-4   11.6µs ± 1%   8.8µs ± 0%  -24.39%  (p=0.000 n=10+8)
ZSortUint64M-4     48.9µs ± 7%  45.1µs ± 6%   -7.75%  (p=0.000 n=10+8)
ZSorterUint64M-4   31.6µs ± 0%  28.8µs ± 0%   -8.96%  (p=0.000 n=9+8)
ZSortUint64L-4     68.2ms ± 0%  62.1ms ± 1%   -8.96%  (p=0.000 n=9+10)
ZSorterUint64L-4   68.4ms ± 0%  61.9ms ± 0%   -9.45%  (p=0.000 n=10+9)
ZSortSortedL-4     2.47ms ± 1%  2.41ms ± 1%   -2.51%  (p=0.000 n=10+10)
ZSorterSortedL-4   1.72ms ± 1%  1.66ms ± 1%   -3.32%  (p=0.000 n=10+9)
```