@wmark 2019-12-11
Floats and *signed integers* can be optimized analogously.