@shawnsmithdev 2019-12-13
That's pretty cool!  I am going to look at this more closely later when I get some time and port those changes to the other types as mentioned... I don't see any obvious reason to reject this.