@wmark 2019-12-13
I've more here (ignore the descents #{[N14]}/2, I'm sorting arrays [N][14]32bit), albeit my “numbers” avail themselves to a different key function as they are not random: https://github.com/wmark/semver/blob/25b609070150e42924ff7c62bf70c0f4fe6e1744/sort.go#L77-L288

I'll try to smooth your keyFn tomorrow, so no hurries.