@wmark 2020-01-03
There's one more low hanging fruit, after you've merged this. Just an outlook:

Building the histogram in one go. Unfortunately due to how and where Go places checks, that would need to be implemented in Assembly. And because Go won't inline anything written in Asm, the entire loop needs to be extracted. Which incidentally will make an succinct implementation.